\name{SDIntroToR-package}
\alias{SDIntroToR-package}
\alias{SDIntroToR}
\docType{package}
\title{
Container for R Introductory Material
}
\description{
This package contains an interactive slide deck meant as a 4-5 hour introduction to R concepts.  RStudio is also introduced as the interface for R.  
}
\details{
\tabular{ll}{
Package: \tab SDIntroToR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-06-15\cr
License: \tab MIT\cr
}
library(SDIntroToR)
openVignette()

}
\author{
Maintainer: Sean Davis <sdavis2@mail.nih.gov>
}
\keyword{ package }
